#Damage Efeect
데미지 이펙트추가,
몬스터 충돌처리 추가.
스킬대미지 및 스텟값 셋팅

+++ MapleStory 2008/MapleStory/Item.cpp
@@ -31,10 +31,10 @@ ITEM CItem::GetItem(void)
 RECT CItem::GetRect(void)
 {
 	RECT rc = {
-		m_tInfo.fX - m_tInfo.fCX / 2.f,
-		m_tInfo.fY - m_tInfo.fCY / 2.f,
-		m_tInfo.fX + m_tInfo.fCX / 2.f,
-		m_tInfo.fY + m_tInfo.fCY / 2.f
+		long(m_tInfo.fX - m_tInfo.fCX / 2.f),
+		long(m_tInfo.fY - m_tInfo.fCY / 2.f),
+		long(m_tInfo.fX + m_tInfo.fCX / 2.f),
+		long(m_tInfo.fY + m_tInfo.fCY / 2.f)
 	};
 
 	return rc;
